Move WAV file type from speech.py to media.py
